# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers (admin)
      security:
        - adminToken: []
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerWithStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer with their orders (admin)
      security:
        - adminToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer with stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerWithStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer info (admin)
      security:
        - adminToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerWithStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer (admin)
      security:
        - adminToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Discounts ────────────────────────────────────────────────────────────────
  /admin/discounts:
    get:
      operationId: listDiscounts
      tags: [discounts]
      summary: List all discount codes (admin)
      security:
        - adminToken: []
      responses:
        "200":
          description: List of discounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Discount"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createDiscount
      tags: [discounts]
      summary: Create a discount code (admin)
      security:
        - adminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscountInput"
      responses:
        "201":
          description: Created discount
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discount"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/discounts/{id}:
    patch:
      operationId: updateDiscount
      tags: [discounts]
      summary: Update a discount code (admin)
      security:
        - adminToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscountUpdate"
      responses:
        "200":
          description: Updated discount
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discount"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteDiscount
      tags: [discounts]
      summary: Delete a discount code (admin)
      security:
        - adminToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /discounts/validate:
    post:
      operationId: validateDiscount
      tags: [discounts]
      summary: Validate a coupon code (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscountValidateInput"
      responses:
        "200":
          description: Discount details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discount"
        "400":
          description: Invalid or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Reviews ──────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    adminToken:
      type: http
      scheme: bearer
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    CustomerWithStats:
      type: object
      required: [id, name, email, phone, totalOrders, totalSpent, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        totalOrders:
          type: integer
        totalSpent:
          type: number
        createdAt:
          type: string
          format: date-time
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          minLength: 1
        phone:
          type: string

    # ── Discount schemas ─────────────────────────────────────────────────────
    Discount:
      type: object
      required: [id, code, type, value, minOrderAmount, maxUses, usedCount, expiresAt, isActive, createdAt]
      properties:
        id:
          type: integer
        code:
          type: string
        type:
          type: string
          enum: [percentage, fixed]
        value:
          type: number
        minOrderAmount:
          type: ["number", "null"]
        maxUses:
          type: ["integer", "null"]
        usedCount:
          type: integer
        expiresAt:
          type: ["string", "null"]
          format: date-time
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    DiscountInput:
      type: object
      required: [code, type, value]
      properties:
        code:
          type: string
          minLength: 1
        type:
          type: string
          enum: [percentage, fixed]
        value:
          type: number
          minimum: 0
        minOrderAmount:
          type: ["number", "null"]
        maxUses:
          type: ["integer", "null"]
        expiresAt:
          type: ["string", "null"]
          format: date-time
        isActive:
          type: boolean
    DiscountUpdate:
      type: object
      properties:
        code:
          type: string
          minLength: 1
        type:
          type: string
          enum: [percentage, fixed]
        value:
          type: number
          minimum: 0
        minOrderAmount:
          type: ["number", "null"]
        maxUses:
          type: ["integer", "null"]
        expiresAt:
          type: ["string", "null"]
          format: date-time
        isActive:
          type: boolean
    DiscountValidateInput:
      type: object
      required: [code, orderTotal]
      properties:
        code:
          type: string
          minLength: 1
        orderTotal:
          type: number
          minimum: 0

    # ── Review schemas ───────────────────────────────────────────────────────
...