# lib/api-zod/src/generated/types/discount.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DiscountType } from './discountType';

export interface Discount {
  id: number;
  code: string;
  type: DiscountType;
  value: number;
  /** @nullable */
  minOrderAmount: number | null;
  /** @nullable */
  maxUses: number | null;
  usedCount: number;
  /** @nullable */
  expiresAt: Date | null;
  isActive: boolean;
  createdAt: Date;
}