# lib/api-zod/src/generated/types/discountUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DiscountUpdateType } from './discountUpdateType';

export interface DiscountUpdate {
  /** @minLength 1 */
  code?: string;
  type?: DiscountUpdateType;
  /** @minimum 0 */
  value?: number;
  /** @nullable */
  minOrderAmount?: number | null;
  /** @nullable */
  maxUses?: number | null;
  /** @nullable */
  expiresAt?: Date | null;
  isActive?: boolean;
}