# lib/api-zod/src/generated/types/discountValidateInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DiscountValidateInput {
  /** @minLength 1 */
  code: string;
  /** @minimum 0 */
  orderTotal: number;
}